+++ tests/test_combine.py
@@ -4,7 +4,7 @@ import pytest
 from ocdsextensionregistry import ProfileBuilder
 from ocdsmerge.exceptions import DuplicateIdValueWarning, InconsistentTypeError
 
-from ocdskit.combine import compile_release_packages, merge, package_records
+from ocdskit.combine import merge, package_records
 from ocdskit.exceptions import (
     InconsistentVersionError,
     MergeErrorWarning,
@@ -141,12 +141,3 @@ def test_merge_inconsistent_type_convert_exceptions_to_warnings(return_package, 
     assert output == expected
     assert len(records) == 1
     assert str(records[0].message) == "ocds-213czf-1: An earlier release had the literal 1 for /integer, but the current release has an object with a 'object' key"  # noqa: E501
-
-
-def test_compile_release_packages():
-    with pytest.warns(DeprecationWarning) as records:
-        compiled_releases = list(compile_release_packages([]))
-
-    assert compiled_releases == []
-    assert len(records) == 1
-    assert str(records[0].message) == 'compile_release_packages() is deprecated. Use merge() instead.'